Include line numbers for strings like i18n/en.yaml when extracting
Right now KDE website strings that come from i18n/en.yaml always end up
with 0 as the line number.

It turns out it's possible to augment PyYAML to include the line number
information as a part of the output. By subclassing str to add just one
property, other things treat it as a normal string while the extraction
process can copy the line number to the POT file.

This commit also includes the key in i18n/en.yaml as context, as I find
it actually useful when translating.